move delete logic to new file
